dbt/requirements.txt
dbt-snowflake==1.7.1
apache-airflow==2.10.3